21-04-25 UserDetail 코드 정리
