+++ src/component/client/UserDetailGame.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const UserDetailGame = () => {
+
+}
+
+export default UserDetailGame
